fix(synthetic-shadow): node-ownership helps to keep the imports order
