+++ packages/@lwc/synthetic-shadow/src/faux-shadow/node.ts
@@ -43,13 +43,9 @@ import { getTextContent } from '../3rdparty/polymer/text-content';
 import { getShadowRoot, isHostElement, getIE11FakeShadowRootPlaceholder } from './shadow-root';
 import { createStaticNodeList } from '../shared/static-node-list';
 import { isGlobalPatchingSkipped } from '../shared/utils';
+import { getNodeNearestOwnerKey, getNodeOwnerKey, isNodeShadowed } from '../shared/node-ownership';
 
 const { DISABLE_NODE_PATCH } = getInitializedFeatureFlags();
-
-// DO NOT CHANGE this:
-// these two values need to be in sync with engine
-const OwnerKey = '$$OwnerKey$$';
-const OwnKey = '$$OwnKey$$';
 
 export const hasNativeSymbolsSupport = Symbol('x').toString() === 'Symbol(x)';
 
@@ -63,56 +59,6 @@ function getInitializedFeatureFlags() {
     }
 
     return { DISABLE_NODE_PATCH };
-}
-
-export function getNodeOwnerKey(node: Node): number | undefined {
-    return node[OwnerKey];
-}
-
-export function setNodeOwnerKey(node: Node, value: number) {
-    if (process.env.NODE_ENV !== 'production') {
-        // in dev-mode, we are more restrictive about what you can do with the owner key
-        defineProperty(node, OwnerKey, {
-            value,
-            configurable: true,
-        });
-    } else {
-        // in prod, for better perf, we just let it roll
-        node[OwnerKey] = value;
-    }
-}
-
-export function setNodeKey(node: Node, value: number) {
-    if (process.env.NODE_ENV !== 'production') {
-        // in dev-mode, we are more restrictive about what you can do with the own key
-        defineProperty(node, OwnKey, {
-            value, // can't be mutated
-        });
-    } else {
-        // in prod, for better perf, we just let it roll
-        node[OwnKey] = value;
-    }
-}
-
-export function getNodeNearestOwnerKey(node: Node): number | undefined {
-    let ownerNode: Node | null = node;
-    let ownerKey: number | undefined;
-    // search for the first element with owner identity (just in case of manually inserted elements)
-    while (!isNull(ownerNode)) {
-        ownerKey = ownerNode[OwnerKey];
-        if (!isUndefined(ownerKey)) {
-            return ownerKey;
-        }
-        ownerNode = parentNodeGetter.call(ownerNode);
-    }
-}
-
-export function getNodeKey(node: Node): number | undefined {
-    return node[OwnKey];
-}
-
-export function isNodeShadowed(node: Node): boolean {
-    return !isUndefined(getNodeNearestOwnerKey(node));
 }
 
 /**
