+++ packages/@lwc/synthetic-shadow/src/faux-shadow/shadow-root.ts
@@ -43,7 +43,8 @@ import { createStaticHTMLCollection } from '../shared/static-html-collection';
 import { getOuterHTML } from '../3rdparty/polymer/outer-html';
 import { retarget } from '../3rdparty/polymer/retarget';
 import { pathComposer } from '../3rdparty/polymer/path-composer';
-import { getInternalChildNodes, setNodeKey, setNodeOwnerKey } from './node';
+import { getInternalChildNodes } from './node';
+import { setNodeKey, setNodeOwnerKey } from '../shared/node-ownership';
 import { innerHTMLSetter } from '../env/element';
 import { getOwnerDocument } from '../shared/utils';
