+++ packages/@lwc/synthetic-shadow/src/index.ts
@@ -15,15 +15,15 @@ import './env/mutation-observer';
 
 // Initialization Routines
 import './polyfills/HTMLSlotElement/main';
-import './polyfills/document-shadow/main';
-import './polyfills/shadow-root/main';
-import './polyfills/event-listener/main';
 import './polyfills/click-event-composed/main'; // must come before event-composed
 import './polyfills/event-composed/main';
 import './polyfills/custom-event-composed/main';
 import './polyfills/focus-event-composed/main';
 import './polyfills/iframe-content-window/main';
 import './polyfills/mutation-observer/main';
+import './polyfills/event-listener/main';
+import './polyfills/document-shadow/main';
+import './polyfills/shadow-root/main';
 
 // Internal Patches
 import './faux-shadow/node';
