+++ packages/@lwc/synthetic-shadow/src/polyfills/document-shadow/polyfill.ts
@@ -23,7 +23,7 @@ import {
     getElementsByTagNameNS as documentGetElementsByTagNameNS,
     querySelectorAll as documentQuerySelectorAll,
 } from '../../env/document';
-import { getNodeOwnerKey } from '../../faux-shadow/node';
+import { getNodeOwnerKey } from '../../shared/node-ownership';
 import { parentElementGetter } from '../../env/node';
 import { retarget } from '../../3rdparty/polymer/retarget';
 import { pathComposer } from '../../3rdparty/polymer/path-composer';
