+++ packages/@lwc/synthetic-shadow/src/polyfills/iframe-content-window/patch.ts
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/MIT
  */
 import { getOwnPropertyDescriptor, defineProperty, isNull, isUndefined } from '@lwc/shared';
-import { getNodeOwnerKey } from '../../faux-shadow/node';
+import { getNodeOwnerKey } from '../../shared/node-ownership';
 
 export default function apply() {
     // the iframe property descriptor for `contentWindow` should always be available, otherwise this method should never be called
